+++ addOns/fuzz/src/main/java/org/zaproxy/zap/extension/fuzz/httpfuzzer/processors/sequence/HttpFuzzerMessageSequenceProcessorUIHandler.java
@@ -1,0 +1,76 @@
+/*
+ * Zed Attack Proxy (ZAP) and its related class files.
+ *
+ * ZAP is an HTTP/HTTPS proxy for assessing web application security.
+ *
+ * Copyright 2015 The ZAP Development Team
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.zaproxy.zap.extension.fuzz.httpfuzzer.processors.sequence;
+
+import org.parosproxy.paros.network.HttpMessage;
+import org.zaproxy.zap.extension.fuzz.httpfuzzer.HttpFuzzerMessageProcessor;
+import org.zaproxy.zap.extension.fuzz.httpfuzzer.HttpFuzzerMessageProcessorUI;
+import org.zaproxy.zap.extension.fuzz.httpfuzzer.HttpFuzzerMessageProcessorUIHandler;
+import org.zaproxy.zap.extension.script.SequenceScript;
+
+public class HttpFuzzerMessageSequenceProcessorUIHandler implements
+        HttpFuzzerMessageProcessorUIHandler<HttpFuzzerMessageSequenceProcessor, HttpFuzzerMessageSequenceProcessorUI> {
+
+    private SequenceScript script;
+
+    public HttpFuzzerMessageSequenceProcessorUIHandler(SequenceScript script) {
+        this.script = script;
+    }
+
+    @Override
+    public String getName() {
+        return HttpFuzzerMessageSequenceProcessor.NAME;
+    }
+
+    @Override
+    public boolean isEnabled(HttpMessage message) {
+        return true;
+    }
+
+    @Override
+    public boolean isDefault() {
+        return true;
+    }
+
+    @Override
+    public HttpFuzzerMessageSequenceProcessorUI createDefault() {
+        return new HttpFuzzerMessageSequenceProcessorUI(script);
+    }
+
+    @Override
+    public Class<HttpMessage> getMessageType() {
+        return HttpMessage.class;
+    }
+
+    @Override
+    public Class<HttpFuzzerMessageSequenceProcessor> getFuzzerMessageProcessorType() {
+        return HttpFuzzerMessageSequenceProcessor.class;
+    }
+
+    @Override
+    public Class<HttpFuzzerMessageSequenceProcessorUI> getFuzzerMessageProcessorUIType() {
+        return HttpFuzzerMessageSequenceProcessorUI.class;
+    }
+
+    @Override
+    public HttpFuzzerMessageSequenceProcessorUIPanel createPanel() {
+        return new HttpFuzzerMessageSequenceProcessorUIPanel(script);
+    }
+}