+++ addOns/fuzz/CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+- Sequence Scripts are now fuzzable.
 - Enable the extensions for all DB types.
 - Use Monospaced font in payload text areas.
 - Possibility to enforce a random order in the RegexPayloadGenerator.